# rapor-model/src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/simulation/track/putTrack/{trackId}": {
      "put": {
        "tags": [
          "simulation-controller"
        ],
        "operationId": "updateTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "azimuth",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "elevation",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/simulation/track/{addTrack}": {
      "post": {
        "tags": [
          "simulation-controller"
        ],
        "operationId": "registerNewTrack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/simulation/track/state": {
      "post": {
        "tags": [
          "simulation-controller"
        ],
        "operationId": "setSimulationState",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/simulation/video/video/{lineID}": {
      "get": {
        "tags": [
          "video-controller"
        ],
        "operationId": "getPPILine",
        "parameters": [
          {
            "name": "lineID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PPILineModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulation/track/tracks": {
      "get": {
        "tags": [
          "simulation-controller"
        ],
        "operationId": "getTracks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulation/track/delTrack/{trackId}": {
      "delete": {
        "tags": [
          "simulation-controller"
        ],
        "operationId": "deleteTrack",
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TrackModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "azimuth": {
            "type": "number",
            "format": "double"
          },
          "range": {
            "type": "number",
            "format": "double"
          },
          "elevation": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PPILineModel": {
        "type": "object",
        "properties": {
          "lineId": {
            "type": "integer",
            "format": "int32"
          },
          "rgbArray": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    }
  }
}